cleanup: Remove dead code
**What**

THe legacy naming of driver and associated neuron components
was `neuron` which was rebranded to `neuronx` a few years ago.
With the automated uploads it will ignore these legacy files and
thus they stay in the repo despite not being useful. Therefore
we need to remove them by hand for the new files that had this
legacy naming scheme.